requirements.txt
Flask==3.0.0
scikit-learn==1.3.2
numpy==1.26.2
pyserial==3.5
